Added check to confirm whether temp tables exist before dropping them at the end of the sprocs. In case of error they may not have been created so there would be another error trying to drop non-existent objects after the catch block.

Related work items: #27729

+++ Icon/Mammoth/MammothDb/Mammoth.Db/eplum/Stored Procedures/GetEPlumAttributes.sql
@@ -753,10 +753,14 @@ BEGIN CATCH
 	THROW
 END CATCH
 
-DROP TABLE #regularPriceKeys
-DROP TABLE #itemExtended
-DROP TABLE #prices
-DROP TABLE #linkedScanCodePriceKeys
+IF OBJECT_ID('tempdb..#regularPriceKeys') IS NOT NULL
+	DROP TABLE #regularPriceKeys
+IF OBJECT_ID('tempdb..#itemExtended') IS NOT NULL
+	DROP TABLE #itemExtended
+IF OBJECT_ID('tempdb..#prices') IS NOT NULL
+	DROP TABLE #prices
+IF OBJECT_ID('tempdb..#linkedScanCodePriceKeys') IS NOT NULL
+	DROP TABLE #linkedScanCodePriceKeys
 
 END
 GO
